Week 1 Assignment
The week 1 assignment was focused on the using the dplyr data package to manipulate and gain insights on a brand new data set for analysis. The Chicago pollution data was a walk through of that process and the assignment I was tasked with creating using the dplyr commands from R Studio to manipulate data and then write a report about those manipulations.